- fix problem with total radiation number changing when moving the house - fix problem with loss of precision of neighboring wall edit points

+++ src/main/java/org/concord/energy3d/model/Foundation.java
@@ -524,7 +524,7 @@ public class Foundation extends HousePart {
 		updateTextureAndColor(mesh, Scene.getInstance().getFoundationColor());
 	}
 
-	public void setSolarValue(final long solarValue) {
+	public void setSolarValue(final double solarValue) {
 		if (solarValue == 0)
 			solarLabel.setVisible(false);
 		else {
